Initial Dev for TD Bootup, Generation, Shutdown
This push includes the initial changes that allow the system to open up a TD instance, generate art, save the art, and shutdown TD. There still needs to be a control mechanism developed that allows the Flask module to start these processes + returns / handles the generated art when TD is complete. Additionally, we need the actual patches / know how to use data to modify the TD patches. This code should definitely be revisited for OOP Design / Overall workflow.